Extract PipelineStep to prevent circular imports
