+++ src/roiextract/pipeline/__init__.py
@@ -1,7 +1,15 @@
-from roiextract.pipeline.pipeline import PipelineStep, ExtractionPipeline
+from roiextract.pipeline.step import PipelineStep
 
 from roiextract.pipeline.inverse import Inverse
-from roiextract.pipeline.roi_aggregation import MeanAggregation
+from roiextract.pipeline.roi_aggregation import MeanAggregation, CentroidAggregation
+
+from roiextract.pipeline.pipeline import ExtractionPipeline
 
 
-__all__ = ["PipelineStep", "ExtractionPipeline", "Inverse", "MeanAggregation"]
+__all__ = [
+    "PipelineStep",
+    "ExtractionPipeline",
+    "Inverse",
+    "MeanAggregation",
+    "CentroidAggregation",
+]
